requirements.txt
python-telegram-bot==21.6
ollama==0.4.7
apscheduler==3.11.2
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
PyYAML>=6.0,<7.0
aiosqlite>=0.20.0,<1.0
structlog==25.1.0
eval_type_backport>=0.2.0
httpx>=0.27.0,<1.0
